оптимизация структуры проекта
