+++ src/page_objects_API/api_cart.py
@@ -1,16 +1,18 @@
+
 class ApiCart:
     def __init__(self, client):
         self.client = client
 
+
     def get_item_cart(self):
-        response = self.client.session.post(f"{self.client.base_url_api}api/sale/cart")
+        response = self.client.session.get(f"{self.client.base_url_api}api/sale/cart")
         return response
 
     def post_cart_add(self, product_id: str, quantity:str):
         data = {'product_id':product_id,
                 'quantity': quantity}
         response = self.client.session.post(f"{self.client.base_url_api}api/sale/cart.add",
-                                           data=data)
+                                            data=data)
         return response
 
 
@@ -18,13 +20,13 @@ class ApiCart:
         data = {'key': key_id,
                 'quantity': quantity}
         response = self.client.session.post(f"{self.client.base_url_api}api/sale/cart.edit",
-                                           data=data)
+                                            data=data)
         return response
 
     def post_cart_remove(self, key_id: str):
         data = {'key': key_id}
         response = self.client.session.post(f"{self.client.base_url_api}api/sale/cart.remove",
-                                           data=data)
+                                            data=data)
         return response
 
     def find_cart_id_of_product_id(self, client, product_id):
