+++ src/page_objects_API/api_payment_address.py
@@ -1,3 +1,6 @@
+import conftest
+
+
 class ApiPayAddress:
     def __init__(self, client):
         self.client = client
